app/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pymango[srv]
pandas
openpyxl
python-jose[cryptography]